+++ resources/js/bootstrap_public.js
@@ -1,7 +1,9 @@
 // Import all of Bootstrap's JS
 import * as bootstrap from 'bootstrap';
 import { Html5Qrcode } from 'html5-qrcode';
+import Sortable from 'sortablejs';
 
 // Expose libraries to the global window object
 window.bootstrap = bootstrap;
 window.Html5Qrcode = Html5Qrcode;
+window.Sortable = Sortable;