+++ resources/views/guest/track.blade.php
@@ -12,7 +12,7 @@
 <body>
     <div class="container-lg mt-5 mb-5">
         <div class="text-center mb-4">
-            <img src="https://upload.wikimedia.org/wikipedia/commons/thumb/3/33/DepEd_logo.svg/1200px-DepEd_logo.svg.png" alt="DepEd Logo" style="height: 80px;">
+            <img src="{{ asset('images/logoipsum-411.png') }}" alt="DepEd Logo" style="height: 80px;">
         </div>
 
         <div id="tracked-documents-container">